requirements.txt
discord.py
flask
python-dotenv
PyNaCl
dnspython
bs4
datetime
dropbox
asyncio
apscheduler
selenium
requests
googletrans==3.1.0a0
requests
liquipediapy
youtube_dl
ffmpeg
fake_useragent
csv